requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
httpx
pydantic
python-jose[cryptography]
python-multipart
openpyxl
apscheduler
python-dotenv
bcrypt

